// src/generated/types/jobResults.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Datto RMM API
 * Datto RMM (Remote Monitoring and Management) REST API
 * OpenAPI spec version: 2.0.0 (Build 201_aadf6556fb176a75fbb9b2aad0f0d2301d64c88a)
 */
import type { JobResultsJobDeploymentStatus } from './jobResultsJobDeploymentStatus';
import type { JobComponentResult } from './jobComponentResult';

/**
 * Job results data
 */
export interface JobResults {
  jobUid?: string;
  deviceUid?: string;
  ranOn?: string;
  jobDeploymentStatus?: JobResultsJobDeploymentStatus;
  componentResults?: JobComponentResult[];
}